refactor: extract MergeAgent filename parsing to utility functions
Extract parseFilenameParts and extractAgentName functions from MergeAgent
into src/agent/utils/mergeFileUtils.ts. This removes 85 lines of pure
string manipulation logic from the agent class, improving code organization
and reusability.

Changes:
- Create src/agent/utils/mergeFileUtils.ts with filename parsing utilities
- Update src/agent/utils/index.ts to export new utilities
- Refactor MergeAgent to use imported utility functions
- Maintain all original functionality and JSDoc documentation
